+++ R/tree_points_methods.R
@@ -1,3 +1,9 @@
+#' Classify point cloud according to voronoi polygons.
+#' @description Performs point cloud segmentation according to regions of influence based on voronoi polygons, calculated from a tree map object.
+#' @examples
+#' file = system.file("extdata", "pine.laz", package="TreeLS")
+#' tls = readTLS(file)
+#' @export
 trees.voronoi = function(){
 
   func = function(las, xymap){
@@ -14,6 +20,7 @@ trees.voronoi = function(){
     return(las)
   }
 
+  func %<>% setAttribute('tpt_mtd')
   return(func)
 }