+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-#### version 1.0.2
-- `tlsTrees` introduced
+#### version 1.1
+- `tlsTrees` and `pointMetrics` introduced
 - inputs changed for `stemPoints`
 - new methods added for usage within `stemPoints` and `stemSegmentation` 
-- minor bug fixes
+- `tlsAlter` name changed to `tlsTransform`
+- minor bug fixes and performance enhancements
 
 #### version 1.0.1
 - bug fix #3